refactor: reorder files, introduce new apistore, driver, constants
